Hooray. The PlantDetailsViewModel looks *fine* and all elements are wired up correctly to get passed into the database. At this point, we're gonna need to start ensuring that other portions update perfectly, and finding a way for notifications to get set satisfactorily.

+++ FloraSaver/Models/IPlant.cs
@@ -12,16 +12,16 @@
         string GivenName { get; set; }
         int Id { get; set; }
         string ImageLocation { get; set; }
-        double? MistInterval { get; set; }
+        int? MistInterval { get; set; }
         string PlantSpecies { get; set; }
-        double? SunInterval { get; set; }
+        int? SunInterval { get; set; }
         TimeSpan TimeOfLastMisting { get; set; }
         TimeSpan TimeOfLastMove { get; set; }
         TimeSpan TimeOfLastWatering { get; set; }
         TimeSpan TimeOfNextMisting { get; set; }
         TimeSpan TimeOfNextMove { get; set; }
         TimeSpan TimeOfNextWatering { get; set; }
-        public double? WaterInterval { get; set; }
+        public int? WaterInterval { get; set; }
 
         double TimeToNextAction(DateTime lastTime, DateTime nextTime);
     }
